pyproject.toml
[project]
name = "setrum"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash>=4.1.0",
    "dash-bootstrap-components>=2.0.4",
    "diskcache>=5.6.3",
    "multiprocess>=0.70.19",
    "pandas>=3.0.1",
    "plotly>=6.6.0",
    "psutil>=7.2.2",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "tabulate>=0.10.0",
]
